requirements.txt
Flask
PyJWT
requests
discord.py
